Delete likes

+++ vkclient/src/main/java/com/alexeyyuditsky/vkclient/data/model/CommentsDto.kt
@@ -3,5 +3,5 @@ package com.alexeyyuditsky.vkclient.data.model
 import com.google.gson.annotations.SerializedName
 
 data class CommentsDto(
-    @SerializedName("count") val count: Int
+    @SerializedName("count") val count: Long
 )
