+++ vkclient/src/main/java/com/alexeyyuditsky/vkclient/data/model/RepostsDto.kt
@@ -3,5 +3,5 @@ package com.alexeyyuditsky.vkclient.data.model
 import com.google.gson.annotations.SerializedName
 
 data class RepostsDto(
-    @SerializedName("count") val count: Int
+    @SerializedName("count") val count: Long
 )
